ACL_PyTorch/built-in/audio/LSTM/requirements.txt
torchvision==0.9.1
onnx==1.7.0
Pillow==7.2.0
onnxruntime==1.7.0
kaldiio==2.17.2
pyyaml==6.0
tensorboard==2.11.0
editdistance==0.6.2
